refactor: small issues
