+++ src/cli/commands/generate.command.ts
@@ -1,6 +1,6 @@
 import got from 'got';
 import { Command } from './command.interface.js';
-import { MockServerData } from '../../shared/types/mock-server-data.type.js';
+import { MockServerData } from '../../shared/types/index.js';
 import { TSVOfferGenerator } from '../../shared/libs/offer-generator/index.js';
 import { TSVFileWriter } from '../../shared/libs/file-writer/index.js';
 import { getErrorMessage } from '../../shared/helpers/common.js';
