+++ src/shared/libs/rest/transform/path-transformer.ts
@@ -19,7 +19,7 @@ export class PathTransformer {
     @inject(Component.Logger) private readonly logger: Logger,
     @inject(Component.Config) private readonly config: Config<RestSchema>,
   ) {
-    this.logger.info('PathTranformer created!');
+    this.logger.info('PathTransformer created!');
   }
 
   private hasDefaultImage(value: string) {
